console working

+++ DesignByContract/Account.cs
@@ -12,43 +12,39 @@ namespace DesignByContract
 
         public String name { get; set; }
    
-        public int amount { get; set; }
+        public double amount { get; set; }
 
-        public Account(string name,  int amount)
+        public Account(string name,  double amount)
         {
             this.name = name;
           
             this.amount = amount;
         }
-
-       
-
+   
         public override string ToString()
         {
             return  " name : " + name + "; amount :" + amount.ToString() ;
         }
 
 
-        public void setNewAmount(int amountn)
+        public void setNewAmount(double amountn)
         {
             amount = amount + amountn;
         }
 
-        public void Deposit(int d)
+        public void Deposit(double d)
         {
             Contract.Requires(d >= 0);
             Contract.Ensures(Contract.OldValue(amount) + d == amount);
             amount += d;
         }
 
-        public void Withdraw(int d)
+        public void Withdraw(double d)
         {
             Contract.Requires(d >=0 );
             Contract.Ensures(Contract.OldValue(amount) - d == amount);
             Contract.EnsuresOnThrow<Exception>(Contract.OldValue(amount) == amount);
-
             if (d > amount) throw new Exception();
-
             amount = amount - d;
         }
     }
